+++ src/Synapse/Synapse/Commands/ManagementCommands/SqlDatabase/NewAzureSynapseSqlDatabase.cs
@@ -51,10 +51,8 @@ namespace Microsoft.Azure.Commands.Synapse
         [ValidateNotNullOrEmpty]
         public string Collation { get; set; }
 
-        [Parameter(ParameterSetName = CreateByNameParameterSet, Mandatory = false, HelpMessage = HelpMessages.MaxSizeInBytes)]
-        [Parameter(ParameterSetName = CreateByParentObjectParameterSet, Mandatory = false, HelpMessage = HelpMessages.MaxSizeInBytes)]
-        [ValidateNotNullOrEmpty]
-        public long MaxSizeInBytes { get; set; }
+        [Parameter(Mandatory = false, HelpMessage = HelpMessages.StorageRedundancy)]
+        public PSSqlDatabaseStorageRedundancyType StorageRedundancy { get; set; }
 
         [Parameter(Mandatory = false, HelpMessage = HelpMessages.AsJob)]
         public SwitchParameter AsJob { get; set; }
@@ -94,8 +92,8 @@ namespace Microsoft.Azure.Commands.Synapse
             {
                 case CreateByNameParameterSet:
                 case CreateByParentObjectParameterSet:
-                    createParams.MaxSizeBytes = this.MaxSizeInBytes;
                     createParams.Collation = this.IsParameterBound(c => c.Collation) ? this.Collation : SynapseConstants.DefaultCollation;
+                    createParams.StorageRedundancy = this.IsParameterBound(c => c.StorageRedundancy) ? this.StorageRedundancy.ToString() : null;
                     break;
 
                 default: throw new AzPSInvalidOperationException(string.Format(Resources.InvalidParameterSet, this.ParameterSetName));