+++ src/Synapse/Synapse/Models/PSSqlDatabaseStorageRedundancyType.cs
@@ -1,0 +1,10 @@
+﻿namespace Microsoft.Azure.Commands.Synapse.Models
+{
+    public enum PSSqlDatabaseStorageRedundancyType
+    {
+        Geo,
+        GeoZone,
+        Local,
+        Zone
+    }
+}